delete the self docker environment

+++ miniserver/test/script/down_hosts.sh
@@ -6,13 +6,19 @@ set -eu
 #####################################################################
 
 THIS_DIR=$(dirname $0)
-COMPOSE_DIR="${THIS_DIR}/.."
+TOP_DIR="${THIS_DIR}/.."
+DOCKER_DIR="${TOP_DIR}/dockertest/ubuntu"
+
+if [ ! -d "${DOCKER_DIR}" ] || [ ! -x "${DOCKER_DIR}" ]; then
+  echo "${0##*/}: <${DOCKER_DIR}> cannot be accessed" 1>&2
+  exit 1
+fi
 
 #####################################################################
 # main routine
 #####################################################################
 
 (
-  cd "${COMPOSE_DIR}"
+  cd "${DOCKER_DIR}"
   docker compose down
 )
